requirements.txt
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
cloudpickle==2.2.0
contourpy==1.0.6
cycler==0.11.0
einops==0.5.0
filelock==3.8.0
fonttools==4.38.0
huggingface-hub==0.10.1
idna==3.4
importlib-resources==5.10.0
IPy==1.1
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
safetensors==0.3.1
setproctitle==1.3.2
six==1.16.0
split-folders==0.4.3
submitit==1.4.5
tensorboardX==2.5.1
thop==0.1.1.post2209072238
timm==0.9.2
torch==1.12.1
torch-summary==1.4.5
torchcam==0.3.2
torchstat==0.0.7
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
